Update to latest changes in JOML

+++ src/org/joml/camera/ScalarMover.java
@@ -24,33 +24,33 @@ package org.joml.camera;
 
 public class ScalarMover {
 
-    public float maxAcceleration = 200.0f;
-    public float maxDeceleration = 200.0f;
-    public float current;
-    public float target;
-    public float velocity;
+    public double maxAcceleration = 200.0f;
+    public double maxDeceleration = 200.0f;
+    public double current;
+    public double target;
+    public double velocity;
 
     public void update(float elapsedTimeInSeconds) {
         if (current == target) {
             return;
         }
-        float currentToTarget = target - current;
-        float directStopDistance = (velocity * velocity) / (2.0f * maxDeceleration);
-        float acceleration = 0.0f;
-        if (velocity * currentToTarget > 0.0f && directStopDistance >= Math.abs(currentToTarget)) {
+        double currentToTarget = target - current;
+        double directStopDistance = (velocity * velocity) / (2.0 * maxDeceleration);
+        double acceleration = 0.0;
+        if (velocity * currentToTarget > 0.0 && directStopDistance >= Math.abs(currentToTarget)) {
             /* Decelerate */
-            float directDec = maxDeceleration;
+            double directDec = maxDeceleration;
             acceleration = (currentToTarget < 0.0 ? -1 : 1) * -directDec;
         } else {
             /* Accelerate */
-            float directAcc = maxAcceleration;
+            double directAcc = maxAcceleration;
             acceleration = (currentToTarget < 0.0 ? -1 : 1) * directAcc;
         }
         velocity += acceleration * elapsedTimeInSeconds;
-        float way = velocity * elapsedTimeInSeconds;
-        if (velocity * currentToTarget > 0.0f && Math.abs(way) > Math.abs(currentToTarget)) {
+        double way = velocity * elapsedTimeInSeconds;
+        if (velocity * currentToTarget > 0.0 && Math.abs(way) > Math.abs(currentToTarget)) {
             /* We would move too far */
-            velocity = 0.0f;
+            velocity = 0.0;
             current = target;
         } else {
             current += way;
